Make PeerMqEngineMoquette more self-contained.
