+++ src/main/resources/WebContent/index.html
@@ -25,15 +25,15 @@
 	  <tr><th>Item</th><th>Status</th></tr>
 	  <tr><td>Uptime</td>
 		<td><div id="uptime"></div></td></tr>
-	  <tr><td>Transfer Messages</td>
+<!--	  <tr><td>Transfer Messages</td>
 		<td><div id="transferedMessages"></div></td></tr>
 	  <tr><td>Storing Messages</td>
-		<td><div id="storedMessages"></div></td></tr>
-	  <tr><td>Storing Topics</td>
+		<td><div id="storedMessages"></div></td></tr> -->
+	  <tr><td>Number of topics</td>
 		<td><div id="storedTopics"></div></td></tr>
-	  <tr><td>Published Messages</td>
+	  <tr><td>Number of published messages</td>
 		<td><div id="publishedMessages"></div></td></tr>
-	  <tr><td>Received Messages from P2P</td>
+	  <tr><td>Number of received messages from P2P</td>
 		<td><div id="receivedMessagesFromP2P"></div></td></tr>
 	</table>
 
@@ -406,9 +406,9 @@
 			console.log("update SUCCESS");
 			console.log(data);
 			if (data != undefined) {
-				$('#storedMessages').text(data.storedMessages);
+//				$('#storedMessages').text(data.storedMessages);
 				$('#storedTopics').text(data.storedTopics);
-				$('#transferedMessages').text(data.transferedMessages);
+//				$('#transferedMessages').text(data.transferedMessages);
 				$('#publishedMessages').text(data.publishedMessages);
 				$('#receivedMessagesFromP2P').text(data.receivedMessagesFromPIAX);
 //				$('#connections').DataTable({
